rust cli: move CLI value parsers out of render
Keep command-line scalar parsing with the CLI definitions so render remains focused on output formatting.
